requirements.txt
mysql-connector-python==9.5.0
pytest==9.0.2
